fix: keep client aborts out of the dev server log
An abort raced against a cold dev server's first request escapes the
request middleware, so h3 prints the 500 itself. Share the detector with
a serve-only plugin that drops the log line.
