+++ src/start.ts
@@ -5,23 +5,11 @@
  */
 import { createCsrfMiddleware, createMiddleware, createStart } from '@tanstack/react-start'
 
+import { isClientAbort } from './lib/clientAbort.ts'
+
 // Server functions are same-origin RPC, so a cross-site page must not drive them with ambient
 // credentials. API routes are exempt: the mobile surface carries its own bearer token.
 const csrfGuard = createCsrfMiddleware({ filter: (ctx) => ctx.handlerType === 'serverFn' })
-
-const ABORT_CODES = new Set(['ECONNRESET', 'ECONNABORTED', 'ERR_STREAM_PREMATURE_CLOSE'])
-
-function isClientAbort(error: unknown): boolean {
-  for (let e: unknown = error, depth = 0; e && depth < 5; depth++) {
-    if (e instanceof Error) {
-      if (e.name === 'AbortError') return true
-      const code = (e as NodeJS.ErrnoException).code
-      if (code && ABORT_CODES.has(code)) return true
-      e = e.cause
-    } else break
-  }
-  return false
-}
 
 // Reloading a page or closing an SSE tab tears the socket down mid-response.
 // Nothing is left to write to, but h3 still treats the abort as an unhandled
